move NoSessionProxy to a separate file
