+++ lib/webdriveragent.js
@@ -5,8 +5,9 @@ import B from 'bluebird';
 import { retryInterval } from 'asyncbox';
 import { SubProcess } from 'teen_process';
 import { JWProxy } from 'appium-base-driver';
-import { fs, logger, util } from 'appium-support';
+import { fs, logger } from 'appium-support';
 import log from './logger';
+import { NoSessionProxy } from "./no-session-proxy";
 import { killAppUsingAppName, generateXcodeConfigFile } from './utils.js';
 import { updateProjectFile, resetProjectFile, checkForDependencies,
          setRealDeviceSecurity, fixXCUICoordinateFile } from './webdriveragent-utils';
@@ -23,37 +24,6 @@ const IPROXY_TIMEOUT = 5000;
 const WDA_AGENT_PORT = 8100;
 const WDA_BASE_URL = 'http://localhost';
 const BUILD_TEST_DELAY = 1000;
-
-
-class NoSessionProxy extends JWProxy {
-  constructor (opts = {}) {
-    super(opts);
-  }
-
-  getUrlForProxy (url) {
-    if (url === '') {
-      url = '/';
-    }
-    const proxyBase = `${this.scheme}://${this.server}:${this.port}${this.base}`;
-    let remainingUrl = '';
-    if ((new RegExp('^/')).test(url)) {
-      remainingUrl = url;
-    } else {
-      throw new Error(`Did not know what to do with url '${url}'`);
-    }
-    remainingUrl = remainingUrl.replace(/\/$/, ''); // can't have trailing slashes
-    return proxyBase + remainingUrl;
-  }
-
-  async proxyReqRes (req, res) {
-    let [response, body] = await this.proxy(req.originalUrl, req.method, req.body);
-    res.headers = response.headers;
-    res.set('Content-type', response.headers['content-type']);
-    body = util.safeJsonParse(body);
-    res.status(response.statusCode).send(JSON.stringify(body));
-  }
-}
-
 
 class WebDriverAgent {
